+++ libft/ft_deg_to_rad.c
@@ -6,16 +6,16 @@
 /*   By: raho <raho@student.hive.fi>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/07 19:52:31 by raho              #+#    #+#             */
-/*   Updated: 2022/12/17 04:13:22 by raho             ###   ########.fr       */
+/*   Updated: 2022/12/21 15:12:11 by raho             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double	ft_deg_to_rad(double degrees)
+float	ft_deg_to_rad(float degrees)
 {
-	double	radians;
+	float	radians;
 
-	radians = (degrees * (double)M_PI) / 180.0;
+	radians = (degrees * (float)M_PI) / 180.0f;
 	return (radians);
 }
